Update test_cve_finding_types.py for correct Severity levels
The RiskLevelSeverity of the two examples used as a test are incorrect. The source (cve.circle.lu) does not offer the Severity level, but according to the CVSS (NVD - Vulnerability Metrics (nist.gov)) this can be determined. Based on this, the correct values are adjusted.
1.	CVE-2021-46882 with CVSS (risk_score) 7.5 should be High (instead of Medium). See also https://nvd.nist.gov/vuln/detail/CVE-2021-46882.
2.	CVE-2016-0616 with CVSS (risk_score) 4.0 should be Medium (instead of Low). See also https://nvd.nist.gov/vuln/detail/CVE-2016-0616

This error has come up after the normalizer has been adjusted (kat_snyk_finding_types & kat_cve_finding_types) with the correct values.